lesson 2, 4, 5
lesson 2:
changed and edited degrees for the code arm motor will be stopped by the robot build. code edited to reset motor position to 0 as the numbers will be off unless reset everytime.
lesson 4:
edited to be correct ports
lesson 5:
changed the example to match current ai instead of including motors and ports in the prompt and choosing correct build instead